Kolibri syscall 6

+++ trunk/include/kunix.h
@@ -25,7 +25,7 @@ errno_t k_stat( const char *path, struct stat *data, int statlink );
 errno_t k_close( int fd );
 
 // Data buffer is malloc'ed and to be freed by caller
-errno_t k_load_file( void **odata, int *osize, const char *fname );
+errno_t k_load_file( void **odata, size_t *osize, const char *fname );
