+++ trunk/oldtree/kernel/phantom/unix/kernel_file.c
@@ -138,7 +138,7 @@ errno_t k_stat( const char *path, struct stat *data, int statlink )
 // -----------------------------------------------------------------------
 
 
-errno_t k_load_file( void **odata, int *osize, const char *fname )
+errno_t k_load_file( void **odata, size_t *osize, const char *fname )
 {
     errno_t e;
     int fd;
